Layer gss_sign() on top of gss_get_mic(), gss_verify() on top of gss_verify_mic(), rather than the other way around. Mechanisms should export a V2 interface.
git-svn-id: svn://anonsvn.mit.edu/krb5/branches/mskrb-integ@21664 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/gssapi/krb5/seal.c
@@ -41,7 +41,7 @@ krb5_gss_wrap(minor_status, context_handle, conf_req_flag,
     gss_buffer_t        output_message_buffer;
 {
     return(kg_seal(minor_status, context_handle, conf_req_flag,
-                   (int) qop_req, input_message_buffer, conf_state,
+                   qop_req, input_message_buffer, conf_state,
                    output_message_buffer, KG_TOK_WRAP_MSG));
 }
 
@@ -58,7 +58,8 @@ krb5_gss_wrap_iov(OM_uint32 *minor_status,
     OM_uint32 major_status;
 
     major_status = kg_seal_iov(minor_status, context_handle, conf_req_flag,
-			       qop_req, conf_state, iov, iov_count, KG_TOK_WRAP_MSG);
+			       qop_req, conf_state,
+			       iov, iov_count, KG_TOK_WRAP_MSG);
 
     return major_status;
 }
